+++ packages/alpine/src/index.js
@@ -4,12 +4,13 @@ import { collapse } from "./collapse.js";
 import { combobox } from "./combobox.js";
 import { dropdown } from "./dropdown.js";
 import { dialog } from "./dialog.js";
+import { navbar } from "./navbar.js";
 import { popover } from "./popover.js";
 import { tabs } from "./tabs.js";
 import { toast } from "./toast.js";
 import { tooltip } from "./tooltip.js";
 
-export { accordion, button, collapse, combobox, dialog, dropdown, popover, tabs, toast, tooltip };
+export { accordion, button, collapse, combobox, dialog, dropdown, navbar, popover, tabs, toast, tooltip };
 
 export function boobstrap(Alpine) {
   Alpine.data("bsAccordion", accordion);
@@ -18,6 +19,7 @@ export function boobstrap(Alpine) {
   Alpine.data("bsCombobox", combobox);
   Alpine.data("bsDropdown", dropdown);
   Alpine.data("bsDialog", dialog);
+  Alpine.data("bsNavbar", navbar);
   Alpine.data("bsPopover", popover);
   Alpine.data("bsTabs", tabs);
   Alpine.data("bsToast", toast);