+++ packages/react/src/index.js
@@ -4,6 +4,7 @@ export { useCollapse } from "./collapse.js";
 export { useCombobox } from "./combobox.js";
 export { useDropdown } from "./dropdown.js";
 export { useDialog } from "./dialog.js";
+export { useNavbar } from "./navbar.js";
 export { usePopover } from "./popover.js";
 export { useTabs } from "./tabs.js";
 export { useToast } from "./toast.js";